+++ index.html
@@ -59,6 +59,6 @@
         </div>
     </div>
 </main>
-<script src="main.js"></script>
+<script  src="main.js"></script>
 </body>
 </html>
