+++ jest.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  transform: {
+    "^.+\\.jsx?$": "babel-jest"
+  },
+}
