+++ scripts/populateBoard.js
@@ -1,0 +1,5 @@
+
+const populateBoard = (location, player) => {
+  document.getElementById('board').children[location].innerText = player;
+}
+export default populateBoard;
